+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 
-## UNRELEASED
+## 1.0.0
 
-*UNRELEASED*
+*2026-05-25*
 
 * Added `--version` CLI option to display current version.
 * Added support for `exclude-newer` in `pixi.devenv.toml`: the option propagates to downstream projects, with the most-downstream value winning.